+++ finagle-core/src/main/scala/com/twitter/finagle/ssl/OpportunisticTls.scala
@@ -1,15 +1,8 @@
 package com.twitter.finagle.ssl
 
-import com.twitter.finagle.Stack
 import com.twitter.io.Buf
 
 object OpportunisticTls {
-
-  case class Param(level: Level)
-
-  object Param {
-    implicit val param = Stack.Param(Param(Off))
-  }
 
   /**
    * Configures the level of TLS that the client or server can support or must