+++ finagle-core/src/main/scala/com/twitter/finagle/ssl/SnoopingLevelInterpreter.scala
@@ -1,18 +1,35 @@
 package com.twitter.finagle.ssl
 
 import com.twitter.finagle.Stack
+import com.twitter.finagle.param.OppTls
 import com.twitter.finagle.ssl.server.SslServerConfiguration
 import com.twitter.finagle.transport.Transport
 import com.twitter.logging.Logger
 
-private[finagle] object SnoopingLevelInterpreter {
+/**
+ * Stack params for configuring TLS snooping
+ *
+ * TLS snooping is a server side feature which allows servers to use both cleartext
+ * and TLS socket connections on the same port. This should only be used if it is
+ * desirable but not required that traffic be encrypted.
+ *
+ * There are two classes of protocols considered: protocols that support TLS
+ * negotiation such as Mux and those that don't such as the framed thrift transport.
+ * The two classes of protocols have a different truth matrix for determining
+ * whether to enable TLS negotiation (see below). For example, a negotiating TLS
+ * connection that _requires_ TLS may enable snooping because if the connection
+ * starts cleartext the session itself may upgrade to a secure connection later.
+ * In contrast, a cleartext HTTP connection has no way to upgrade to a secure
+ * connection and thus if TLS is required snooping must be disabled.
+ */
+object SnoopingLevelInterpreter {
 
   private val logger = Logger.get()
 
   case class Param(interpreter: Interpreter)
 
   object Param {
-    implicit lazy val param: Stack.Param[Param] = Stack.Param(Param(Disabled))
+    implicit lazy val param: Stack.Param[Param] = Stack.Param(Off)
   }
 
   sealed abstract class Interpreter
@@ -23,6 +40,11 @@ private[finagle] object SnoopingLevelInterpreter {
   /** TLS snooping is enabled based on the predicate function */
   case class Enabled(predicate: (OpportunisticTls.Level, SslServerConfiguration) => Boolean)
       extends Interpreter
+
+  /**
+   * Tls snooping configuration that disables TLS snooping.
+   */
+  val Off: Param = Param(Disabled)
 
   /**
    * TLS Snooping configuration for protocols that don't support negotiation.
@@ -74,7 +96,8 @@ private[finagle] object SnoopingLevelInterpreter {
     params[SnoopingLevelInterpreter.Param].interpreter match {
       case Disabled => false
       case Enabled(enableSnooping) =>
-        val level = params[OpportunisticTls.Param].level
+        // If the opportunistic level is not configured we default to desired.
+        val level = params[OppTls].level.getOrElse(OpportunisticTls.Desired)
         params[Transport.ServerSsl].sslServerConfiguration match {
           case Some(config) =>
             enableSnooping(level, config)