+++ finagle-http/src/test/scala/com/twitter/finagle/http/AlpnToTlsSnoopingEndToEndTest.scala
@@ -2,6 +2,7 @@ package com.twitter.finagle.http
 
 import com.twitter.finagle.Http
 import com.twitter.finagle.http.ssl.HttpSslTestComponents
+import com.twitter.finagle.param.OppTls
 import com.twitter.finagle.ssl.OpportunisticTls
 import com.twitter.finagle.ssl.server.SslServerConfiguration
 
@@ -14,5 +15,5 @@ class AlpnToTlsSnoopingEndToEndTest extends Http2AlpnTest {
   override def serverImpl(): Http.Server =
     super
       .serverImpl()
-      .configured(OpportunisticTls.Param(OpportunisticTls.Desired))
+      .configured(OppTls(Some(OpportunisticTls.Desired)))
 }