+++ finagle-thriftmux/src/test/scala/com/twitter/finagle/thriftmux/pushsession/MuxDowngradingNegotiatorTest.scala
@@ -1,22 +1,26 @@
 package com.twitter.finagle.thriftmux.pushsession
 
 import com.twitter.conversions.DurationOps._
-import com.twitter.finagle.Mux.param.OppTls
 import com.twitter.finagle.pushsession.RefPushSession
 import com.twitter.finagle.pushsession.utils.MockChannelHandle
 import com.twitter.finagle.{param => fparam}
 import com.twitter.finagle.mux.{Request, Response}
-import com.twitter.finagle.mux.transport.{Message, OpportunisticTls}
+import com.twitter.finagle.mux.transport.Message
 import com.twitter.finagle.{Service, Stack, ThriftMux}
 import com.twitter.finagle.mux.pushsession.{
   MuxChannelHandle,
   MuxClientNegotiatingSession,
   SharedNegotiationStats
 }
+import com.twitter.finagle.param.OppTls
 import com.twitter.finagle.stats.InMemoryStatsReceiver
+import com.twitter.finagle.ssl.OpportunisticTls
+import com.twitter.finagle.ssl.session.SslSessionInfo
 import com.twitter.finagle.thriftmux.thriftscala.TestService
 import com.twitter.io.{Buf, ByteReader}
-import com.twitter.util.Future
+import com.twitter.util.{Future, Try}
+import java.security.cert.X509Certificate
+import javax.net.ssl.SSLSession
 import org.apache.thrift.protocol.TBinaryProtocol
 import org.apache.thrift.transport.{TFramedTransport, TMemoryBuffer}
 import org.scalatest.FunSuite
@@ -24,13 +28,31 @@ import org.scalatest.FunSuite
 class MuxDowngradingNegotiatorTest extends FunSuite {
 
   private class Ctx {
-    private class MockMuxChannelHandle
+    class MockMuxChannelHandle
         extends MuxChannelHandle(
           underlying = handle,
           ch = null, // only used in overridden method `sendAndForgetNow`
           params = params
         ) {
       override def sendNowAndForget(buf: Buf): Unit = sendAndForget(buf)
+    }
+
+    class TlsMockMuxChannelHandle extends MockMuxChannelHandle {
+
+      var tlsActivated: Boolean = false
+
+      override def turnOnTls(onHandshakeComplete: Try[Unit] => Unit): Unit = {
+        tlsActivated = true
+      }
+
+      override val sslSessionInfo: SslSessionInfo = new SslSessionInfo {
+        def usingSsl: Boolean = true
+        def session: SSLSession = ???
+        def sessionId: String = ???
+        def cipherSuite: String = ???
+        def localCertificates: Seq[X509Certificate] = ???
+        def peerCertificates: Seq[X509Certificate] = ???
+      }
     }
 
     lazy val statsReceiver: InMemoryStatsReceiver = new InMemoryStatsReceiver
@@ -123,6 +145,36 @@ class MuxDowngradingNegotiatorTest extends FunSuite {
     }
   }
 
+  test(
+    "allows downgraded session if opportunistic TLS is required AND the transport is already encrypted") {
+    new Ctx {
+      override lazy val params: Stack.Params =
+        ThriftMux.BaseServerParams + fparam.Stats(statsReceiver) +
+          OppTls(Some(OpportunisticTls.Required))
+
+      override lazy val muxChannelHandle: TlsMockMuxChannelHandle = new TlsMockMuxChannelHandle
+
+      val thriftMsg = {
+        val buffer = new TMemoryBuffer(1)
+        val framed = new TFramedTransport(buffer)
+        val proto = new TBinaryProtocol(framed)
+        val arg = TestService.Query.Args("hi")
+        TestService.Query.Args.encode(arg, proto)
+        framed.flush()
+        val bytes = buffer.getArray
+        Buf.ByteArray.Owned(bytes)
+      }
+
+      refSession.receive(ByteReader(thriftMsg))
+
+      // We don't need to initiate TLS if it's already started
+      assert(!muxChannelHandle.tlsActivated)
+
+      assert(!handle.closedCalled)
+      assert(statsReceiver.counters(Seq("thriftmux", "downgraded_connects")) == 1L)
+    }
+  }
+
   test("no infinite loops if we close before the handshake is complete and fail negotiation") {
     new Ctx {
       override lazy val params: Stack.Params =
@@ -143,7 +195,6 @@ class MuxDowngradingNegotiatorTest extends FunSuite {
       val closeF = refSession.close(60.seconds)
       handle.serialExecutor.executeAll()
       refSession.receive(ByteReader(thriftMsg))
-
       assert(handle.closedCalled)
       handle.onClosePromise.setDone()
       assert(closeF.isDefined)