+++ finagle-thriftmux/src/test/scala/com/twitter/finagle/thriftmux/ssl/VanillaThriftSmuxSslTest.scala
@@ -5,22 +5,26 @@ import com.twitter.finagle.ssl.OpportunisticTls
 import com.twitter.finagle.ssl.client.SslClientSessionVerifier
 import com.twitter.finagle.ssl.server.SslServerSessionVerifier
 import com.twitter.finagle.stats.StatsReceiver
-import com.twitter.finagle.thriftmux.ssl.ThriftSmuxSslTestComponents.{mkTlsClient, mkTlsServer}
+import com.twitter.finagle.thriftmux.ssl.ThriftSmuxSslTestComponents.{
+  mkTlsServer,
+  mkTlsVanillaThriftClient
+}
 import com.twitter.finagle.thriftmux.thriftscala.TestService
 
-class ThriftSmuxSslTest extends AbstractThriftSmuxSslTest {
+// Tests that we can use a vanilla thrift client to talk to a thriftmux server that can do snooping.
+class VanillaThriftSmuxSslTest extends AbstractThriftSmuxSslTest {
   protected def doMkTlsClient(
     port: Int,
     label: String,
     statsReceiver: StatsReceiver,
     sessionVerifier: SslClientSessionVerifier
   ): TestService.MethodPerEndpoint =
-    mkTlsClient(port, label, statsReceiver, sessionVerifier, Some(OpportunisticTls.Required))
+    mkTlsVanillaThriftClient(port, label, statsReceiver, sessionVerifier)
 
   protected def doMkTlsServer(
     label: String,
     statsReceiver: StatsReceiver,
     sessionVerifier: SslServerSessionVerifier
   ): ListeningServer =
-    mkTlsServer(label, statsReceiver, sessionVerifier, false, Some(OpportunisticTls.Required))
+    mkTlsServer(label, statsReceiver, sessionVerifier, true, Some(OpportunisticTls.Required))
 }